+++ gdrive/gdrive-file-contents.h
@@ -173,11 +173,11 @@ size_t gdrive_fcontents_read(Gdrive_File_Contents* pContents,
                              size_t size
 );
 
-size_t gdrive_fcontents_write(Gdrive_File_Contents* pContents, 
-                              const char* buf, 
-                              off_t offset,
-                              size_t size,
-                              bool extendChunk
+off_t gdrive_fcontents_write(Gdrive_File_Contents* pContents, 
+                             const char* buf, 
+                             off_t offset,
+                             size_t size,
+                             bool extendChunk
 );
